started modularising SurveyCard
