+++ Decsys/ClientApp/package.json
@@ -4,12 +4,12 @@
   "private": true,
   "dependencies": {
     "@smooth-ui/core-sc": "^8.1.0",
+    "prop-types": "^15.6.2",
     "react": "^16.7.0",
     "react-dom": "^16.7.0",
     "react-redux": "^6.0.0",
     "react-router-dom": "^4.3.1",
     "react-scripts": "2.1.3",
-    "react-tabulator": "^0.7.11",
     "redux": "^4.0.1",
     "styled-components": "^4.1.3",
     "styled-icons": "^6.6.0"
