+++ Decsys/ClientApp/src/themes/index.js
@@ -2,5 +2,6 @@ import { theme } from "@smooth-ui/core-sc";
 
 export default {
   ...theme,
+  name: "decsys",
   primary: "#007bff"
 };
